+++ scripts/server-unit.test.mjs
@@ -215,6 +215,8 @@ assert.deepEqual(envSettingsStore.validateEnvSettingsForTest({
   APP_SIDE_PANEL_FADE_MS: '1200',
   APP_AMBIENT_MOTION_ENABLED: 'false',
   APP_AMBIENT_MOTION_DURATION_MS: '120000',
+  APP_TOOLTIPS_ENABLED: 'false',
+  APP_TOOLTIP_DELAY_MS: '1200',
   AZURE_CLIENT_SECRET: ''
 }), {
   PORT: '3001',
@@ -224,12 +226,15 @@ assert.deepEqual(envSettingsStore.validateEnvSettingsForTest({
   APP_SIDE_PANEL_IDLE_MS: '30000',
   APP_SIDE_PANEL_FADE_MS: '1200',
   APP_AMBIENT_MOTION_ENABLED: 'false',
-  APP_AMBIENT_MOTION_DURATION_MS: '120000'
+  APP_AMBIENT_MOTION_DURATION_MS: '120000',
+  APP_TOOLTIPS_ENABLED: 'false',
+  APP_TOOLTIP_DELAY_MS: '1200'
 });
 assert.throws(() => envSettingsStore.validateEnvSettingsForTest({ PORT: '99999' }), /at most 65535/);
 assert.throws(() => envSettingsStore.validateEnvSettingsForTest({ NODE_ENV: 'staging' }), /must be one of/);
 assert.throws(() => envSettingsStore.validateEnvSettingsForTest({ APP_SIDE_PANEL_IDLE_MS: '500' }), /at least 1000/);
 assert.throws(() => envSettingsStore.validateEnvSettingsForTest({ APP_AMBIENT_MOTION_DURATION_MS: '1000' }), /at least 30000/);
+assert.throws(() => envSettingsStore.validateEnvSettingsForTest({ APP_TOOLTIP_DELAY_MS: '9000' }), /at most 3000/);
 assert.throws(() => envSettingsStore.validateEnvSettingsForTest({ UNKNOWN_SETTING: 'x' }), /Unknown setting/);
 
 const okResponse = await nextHandler.runHandler((_req, res) => res.json({ success: true, value: 42 }), makeReq('http://127.0.0.1:3000/api/unit'));